refactor: move transaction.py into transactions dir
